added long long support

+++ euler.c
@@ -1,7 +1,7 @@
 #include <math.h>
 #include <stdlib.h>
 
-int a(int p, int q) {
+long long a(long long p, long long q) {
   return (
     (pow(q, 2) + p * q - pow(p, 2))
     *
@@ -9,7 +9,7 @@ int a(int p, int q) {
   );
 }
 
-int b(int p, int q) {
+long long b(long long p, long long q) {
   return (
     (pow(p, 2) + pow(q, 2))
     *
@@ -17,7 +17,7 @@ int b(int p, int q) {
   );
 }
 
-int c(int p, int q) {
+long long c(long long p, long long q) {
   return (
     2.0
     *
@@ -27,24 +27,24 @@ int c(int p, int q) {
   );
 }
 
-void print_triples(long *triples) {
+void print_triples(long long *triples) {
 
 }
 
-void print_triple(long *triple) {
+void print_triple(long long *triple) {
 
 }
 
-long* cons_triple(int a, int b, int c) {
-  long *triple = (long*) malloc(3 * sizeof(long*));
+long long* cons_triple(long long a, long long b, long long c) {
+  long long *triple = (long long*) malloc(3 * sizeof(long long*));
   triple[0] = a;
   triple[1] = b;
   triple[2] = c;
   return triple;
 }
 
-long* cons_duo(int a, int b) {
-  long *db = (long*) malloc(2 * sizeof(long*));
+long long* cons_duo(long long a, long long b) {
+  long long *db = (long long*) malloc(2 * sizeof(long long*));
   db[0] = a;
   db[1] = b;
   return db;
